Revert "kernel: inline z_unpend_first_thread()"
This reverts commit f6a76c32b7994464984f1b7fb1ccdbfc4ee8c9b6.
